restructure frontend a bit
