+++ .eslintrc.json
@@ -27,7 +27,8 @@
             "double"
         ],
         "react/jsx-uses-react": "error",
-        "react/jsx-uses-vars": "error"
+        "react/jsx-uses-vars": "error",
+        "react/prop-types": "ignore"
     },
     "settings": {
         "react": {
