+++ app/javascript/application.js
@@ -1,5 +1,4 @@
 // Entry point for the build script in your package.json
-import "@hotwired/turbo-rails";
-import "./components";
+import "./app";
 // eslint-disable-next-line no-unused-vars
 import * as bootstrap from "bootstrap";
