--- app/javascript/controllers/application.js
@@ -1,9 +1,0 @@
-import { Application } from "@hotwired/stimulus";
-
-const application = Application.start();
-
-// Configure Stimulus development experience
-application.debug = false;
-window.Stimulus   = application;
-
-export { application };
