+++ app/javascript/routes/index.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "../components/Home";
+import Home from "../screens/Home";
 
 export default (
   <Router>
